+++ src/frontends/pytorch/src/utils.hpp
@@ -74,7 +74,7 @@ template <typename T>
 OutputVector translate_1to1_match_1_inputs(NodeContext& context) {
     auto inputs = context.inputs();
     FRONT_END_OP_CONVERSION_CHECK(inputs.size() >= 1, "Operation has no inputs.");
-    for (int i = 1; i < inputs.size(); i++) {
+    for (size_t i = 1; i < inputs.size(); i++) {
         FRONT_END_OP_CONVERSION_CHECK(context.input_is_none(i), "Got more inputs than expected.");
     }
     FRONT_END_OP_CONVERSION_CHECK(!context.input_is_none(0), "Input should not be None.");
@@ -85,7 +85,7 @@ template <typename T>
 OutputVector translate_1to1_match_2_inputs(NodeContext& context) {
     auto inputs = context.inputs();
     FRONT_END_OP_CONVERSION_CHECK(inputs.size() >= 2, "Operation has less then 2 inputs.");
-    for (int i = 2; i < inputs.size(); i++) {
+    for (size_t i = 2; i < inputs.size(); i++) {
         FRONT_END_OP_CONVERSION_CHECK(context.input_is_none(i), "Got more inputs than expected.");
     }
     FRONT_END_OP_CONVERSION_CHECK(!context.input_is_none(0) && !context.input_is_none(1), "Inputs should not be None.");
@@ -96,7 +96,7 @@ template <typename T>
 OutputVector translate_1to1_match_2_inputs_align_types(NodeContext& context) {
     auto inputs = context.inputs();
     FRONT_END_OP_CONVERSION_CHECK(inputs.size() >= 2, "Operation has less then 2 inputs.");
-    for (int i = 2; i < inputs.size(); i++) {
+    for (size_t i = 2; i < inputs.size(); i++) {
         FRONT_END_OP_CONVERSION_CHECK(context.input_is_none(i), "Got more inputs than expected.");
     }
     FRONT_END_OP_CONVERSION_CHECK(!context.input_is_none(0) && !context.input_is_none(1), "Inputs should not be None.");
